pyproject.toml
[tool.poetry]
name = "py-aiger-dfa"
version = "0.3.3"
description = "Library for moving between sequential circuits AIGs and DFAs."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
repository = "https://github.com/mvcisback/py-aiger-dfa"
readme="README.md"
license = "MIT"
packages = [
    { include = "aiger_dfa" },
]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^21.0.0"
bidict = "^0.21.0"
funcy = "^1.12"
dfa = "^4"
py-aiger = "^6.0.0"
py-aiger-bv = "^4.7.4"
py-aiger-ptltl = "^3.0.0"
pyrsistent = "^0.18"


[tool.poetry.dev-dependencies]
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-xdist = "^2.0.0"
coverage = "^5.0.0"
pdoc3 = "^0.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
